Assert dedupe Z-set multiplicity, document rollback lifecycle, add users guide
Movement-dedupe weight oracle (fixes the multiplicity-2 test gap):
- decide dedupe test helper now retains each consolidated row's Z-set weight
  and asserts multiplicity == 1, so a single decision emitted with weight 2
  fails the test. The behavioural `duplicate_targets_produce_single_decision`
  test asserts the same on movement_decision_stream's output.
- Split decide.rs into decide/mod.rs + decide/tests.rs to keep both within the
  module-line budget after retaining the weight.

state.rs: document the frame-rollback lifecycle on `begin_frame_rollback` —
the begin/record-undo/stash/commit-or-rollback sequence with text examples for
both the successful-commit and failed-rollback paths (no behaviour change).

Docs:
- Add docs/users-guide.md documenting user-facing primary-map asset path
  validation (relative-only; rooted/UNC/drive-absolute and `..`-component
  paths rejected; substring `..` accepted), linked from docs/README.md.
- Add a "Testing strategy for invariants" section to docs/developers-guide.md
  explaining the bounded near-exhaustive rstest matrices and the deliberate
  non-adoption of proptest/Kani/Verus (no new dependency).
